Remove remaining internal published API methods from activity store context.
