+++ formula-android/src/main/java/com/instacart/formula/android/StreamConfigurator.kt
@@ -1,0 +1,23 @@
+package com.instacart.formula.android
+
+import androidx.fragment.app.FragmentActivity
+import io.reactivex.Observable
+import io.reactivex.disposables.Disposable
+
+/**
+ * Provides ability to configure RxJava streams that will survive configuration changes.
+ */
+interface StreamConfigurator<out Activity : FragmentActivity> {
+
+    /**
+     * Keeps activity in-sync with state observable updates. On activity configuration
+     * changes, the last update is applied to new activity instance.
+     *
+     * @param state a state observable
+     * @param update an update function
+     */
+    fun <State> update(
+        state: Observable<State>,
+        update: (Activity, State) -> Unit
+    ): Disposable
+}
