+++ formula-android/src/main/java/com/instacart/formula/integration/ActivityStore.kt
@@ -1,6 +1,7 @@
 package com.instacart.formula.integration
 
 import androidx.fragment.app.FragmentActivity
+import com.instacart.formula.android.StreamConfigurator
 import com.instacart.formula.fragment.FragmentFlowState
 import com.instacart.formula.fragment.FragmentFlowStore
 import com.instacart.formula.fragment.FragmentLifecycleEvent
@@ -13,7 +14,7 @@ import io.reactivex.disposables.Disposable
  *
  * @param contracts Fragment state management defined for this [Activity].
  * @param streams This provides ability to configure arbitrary RxJava streams that survive
- *                configuration changes. Check [ActivityStoreContext.StreamConfigurator] for utility methods.
+ *                configuration changes. Check [com.instacart.formula.android.StreamConfigurator] for utility methods.
  * @param configureActivity This is invoked as part of [com.instacart.formula.FormulaAndroid.onPreCreate]. You can
  *                          use this callback to inject the activity.
  * @param onRenderFragmentState This is invoked after [FragmentFlowState] has been updated.
@@ -21,7 +22,7 @@ import io.reactivex.disposables.Disposable
  */
 class ActivityStore<Activity : FragmentActivity>(
     val contracts: FragmentFlowStore,
-    val streams: (() -> Disposable)? = null,
+    val streams: (StreamConfigurator<Activity>.() -> Disposable)? = null,
     val configureActivity: ((Activity) -> Unit)? = null,
     val onRenderFragmentState: ((Activity, FragmentFlowState) -> Unit)? = null,
     val onFragmentLifecycleEvent: ((FragmentLifecycleEvent) -> Unit)? = null
