+++ formula-android/src/main/java/com/instacart/formula/integration/ActivityStoreContext.kt
@@ -3,13 +3,14 @@ package com.instacart.formula.integration
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.Lifecycle
 import com.instacart.formula.activity.ActivityResult
+import com.instacart.formula.android.StreamConfigurator
+import com.instacart.formula.android.internal.StreamConfiguratorIml
 import com.instacart.formula.fragment.FragmentContract
 import com.instacart.formula.fragment.FragmentFlowState
 import com.instacart.formula.fragment.FragmentFlowStore
 import com.instacart.formula.fragment.FragmentLifecycleEvent
 import io.reactivex.Observable
 import io.reactivex.disposables.Disposable
-import io.reactivex.functions.BiFunction
 
 /**
  * This class provides context within which you can create [ActivityStore]. It provides
@@ -19,12 +20,6 @@ import io.reactivex.functions.BiFunction
  * @param Activity Type of activity that this class provides context for.
  */
 abstract class ActivityStoreContext<out Activity : FragmentActivity> {
-
-    @PublishedApi internal abstract fun startedActivity(): Activity?
-
-    @PublishedApi internal abstract fun activityAttachEvents(): Observable<Boolean>
-
-    @PublishedApi internal abstract fun activityStartedEvents(): Observable<Unit>
 
     /**
      * Events for [FragmentActivity.onActivityResult].
@@ -54,6 +49,21 @@ abstract class ActivityStoreContext<out Activity : FragmentActivity> {
     abstract fun isFragmentResumed(contract: FragmentContract<*>): Observable<Boolean>
 
     /**
+     * This enables you to select specific events from the activity.
+     *
+     * @param Event Type of event
+     */
+    abstract fun <Event> selectActivityEvents(
+        select: Activity.() -> Observable<Event>
+    ): Observable<Event>
+
+    /**
+     * Performs an [effect] on the current activity instance. If there is no activity connected,
+     * it will do nothing.
+     */
+    abstract fun send(effect: Activity.() -> Unit)
+
+    /**
      * Creates an [ActivityStore].
      *
      * @param configureActivity This is called when activity is created before view inflation. You can use this to
@@ -68,17 +78,15 @@ abstract class ActivityStoreContext<out Activity : FragmentActivity> {
         configureActivity: (ActivityT.() -> Unit)? = null,
         onRenderFragmentState: ((ActivityT, FragmentFlowState) -> Unit)? = null,
         onFragmentLifecycleEvent: ((FragmentLifecycleEvent) -> Unit)? = null,
-        streams: (StreamConfigurator<Activity>.() -> Disposable)? = null,
+        streams: (StreamConfigurator<ActivityT>.() -> Disposable)? = null,
         contracts: FragmentFlowStore
     ): ActivityStore<ActivityT> {
-        val streamStart = streams?.let { streams(it) }
-
         return ActivityStore(
             contracts =  contracts,
             configureActivity = configureActivity,
             onFragmentLifecycleEvent = onFragmentLifecycleEvent,
             onRenderFragmentState = onRenderFragmentState,
-            streams = streamStart
+            streams = streams
         )
     }
 
@@ -97,7 +105,7 @@ abstract class ActivityStoreContext<out Activity : FragmentActivity> {
         noinline configureActivity: (ActivityT.() -> Unit)? = null,
         noinline onRenderFragmentState: ((ActivityT, FragmentFlowState) -> Unit)? = null,
         noinline onFragmentLifecycleEvent: ((FragmentLifecycleEvent) -> Unit)? = null,
-        noinline streams: (StreamConfigurator<Activity>.() -> Disposable)? = null,
+        noinline streams: (StreamConfigurator<ActivityT>.() -> Disposable)? = null,
         crossinline contracts: FragmentBindingBuilder<Unit>.() -> Unit = {}
     ): ActivityStore<ActivityT> {
         return store(
@@ -110,69 +118,12 @@ abstract class ActivityStoreContext<out Activity : FragmentActivity> {
     }
 
     /**
-     * Creates [FragmentFlowStore] with a [Component] instance.
+     * Convenience method to to create a [FragmentFlowStore] with a [Component] instance.
      */
     inline fun <Component> contracts(
         rootComponent: Component,
         crossinline contracts: FragmentBindingBuilder<Component>.() -> Unit
     ): FragmentFlowStore {
         return FragmentFlowStore.init(rootComponent, contracts)
-    }
-
-    /**
-     * Performs an [effect] on the current activity instance. If there is no activity connected,
-     * it will do nothing.
-     */
-    inline fun send(effect: Activity.() -> Unit) {
-        // We allow emitting effects only after activity has started
-        startedActivity()?.effect() ?: run {
-            // Log missing activity.
-        }
-    }
-
-    /**
-     * This enables you to select specific events from the activity.
-     *
-     * @param Event Type of event
-     */
-    abstract fun <Event> selectActivityEvents(
-        select: Activity.() -> Observable<Event>
-    ): Observable<Event>
-
-    private fun streams(configure: StreamConfigurator<Activity>.() -> Disposable): () -> Disposable {
-        return { StreamConfigurator(this).configure() }
-    }
-
-    /**
-     * Provides ability to configure RxJava streams that will survive configuration changes.
-     */
-    class StreamConfigurator<out Activity : FragmentActivity>(
-        private val context: ActivityStoreContext<Activity>
-    ) {
-        /**
-         * Keeps activity in-sync with state observable updates. On activity configuration
-         * changes, the last update is applied to new activity instance.
-         *
-         * @param state a state observable
-         * @param update an update function
-         */
-        fun <State> update(
-            state: Observable<State>,
-            update: (Activity, State) -> Unit
-        ): Disposable {
-            // To keep activity & state in sync, we re-emit state on every activity change.
-            val stateEmissions = Observable.combineLatest(
-                state,
-                context.activityStartedEvents(),
-                BiFunction<State, Unit, State> { state, event ->
-                    state
-                }
-            )
-            return stateEmissions.subscribe { state ->
-                context.startedActivity()?.let {
-                    update(it, state)
-                }
-            }
-        }
     }
 }
